+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Fix error message for json parse whitespace in `strict`
   * Prevent loss of async hooks context
   * Prevent hanging when request already read
   * deps: depd@2.0.0